reorganize proof input and output
