+++ crates/sovereign-sdk/module-system/sov-modules-api/src/lib.rs
@@ -196,7 +196,8 @@ pub use sov_rollup_interface::soft_confirmation::{
     SignedSoftConfirmation, UnsignedSoftConfirmation, UnsignedSoftConfirmationV1,
 };
 pub use sov_rollup_interface::stf::StateDiff;
-pub use sov_rollup_interface::zk::{BatchProofCircuitOutput, Zkvm};
+pub use sov_rollup_interface::zk::batch_proof::output::v2::BatchProofCircuitOutput;
+pub use sov_rollup_interface::zk::Zkvm;
 pub use sov_rollup_interface::{digest, BasicAddress, RollupAddress};
 
 pub mod prelude {
