+++ crates/sovereign-sdk/rollup-interface/src/state_machine/zk/batch_proof/mod.rs
@@ -1,0 +1,4 @@
+/// Batch proof input types and functions
+pub mod input;
+/// Batch proof output types and functions
+pub mod output;
