+++ crates/sovereign-sdk/rollup-interface/src/state_machine/zk/batch_proof/output/mod.rs
@@ -1,0 +1,9 @@
+use alloc::collections::BTreeMap;
+
+/// Genesis output module
+pub mod v1;
+/// Kumquat output module
+pub mod v2;
+
+/// State diff produced by the Zk proof
+pub type CumulativeStateDiff = BTreeMap<Vec<u8>, Option<Vec<u8>>>;
