+++ crates/sovereign-sdk/rollup-interface/src/state_machine/zk/batch_proof/output/v2.rs
@@ -1,0 +1,48 @@
+use alloc::vec::Vec;
+use core::fmt::Debug;
+
+use borsh::{BorshDeserialize, BorshSerialize};
+use serde::{Deserialize, Serialize};
+
+use super::CumulativeStateDiff;
+use crate::da::DaSpec;
+
+// TODO: use name with v2 now.
+
+/// The public output of a SNARK batch proof in Sovereign, this struct makes a claim that
+/// the state of the rollup has transitioned from `initial_state_root` to `final_state_root`
+///
+/// The period of time covered by a state transition proof is a range of L2 blocks whose sequencer
+/// commitments are included in the DA slot with hash `da_slot_hash`. The range is inclusive.
+/// Some fields (prev_soft_confirmation_hash, final_soft_confirmation_hash and last_l2_height)
+/// Will be 0 for pre fork 1 proofs because this is a new output format and those fields
+/// did not exist pre fork 1
+#[derive(Clone, Debug, Serialize, Deserialize, BorshSerialize, BorshDeserialize, PartialEq, Eq)]
+pub struct BatchProofCircuitOutput<Da: DaSpec, Root> {
+    /// The state of the rollup before the transition
+    pub initial_state_root: Root,
+    /// The state of the rollup after the transition
+    pub final_state_root: Root,
+    /// The hash of the last soft confirmation before the state transition
+    /// This will be [0; 32] for pre fork 1 proofs
+    pub prev_soft_confirmation_hash: [u8; 32],
+    /// The hash of the last soft confirmation in the state transition
+    /// This will be [0; 32] for pre fork 1 proofs
+    pub final_soft_confirmation_hash: [u8; 32],
+    /// State diff of L2 blocks in the processed sequencer commitments.
+    pub state_diff: CumulativeStateDiff,
+    /// The DA slot hash that the sequencer commitments causing this state transition were found in.
+    pub da_slot_hash: Da::SlotHash,
+    /// The range of sequencer commitments in the DA slot that were processed.
+    /// The range is inclusive.
+    pub sequencer_commitments_range: (u32, u32),
+    /// Sequencer public key.
+    pub sequencer_public_key: Vec<u8>,
+    /// Sequencer DA public key.
+    pub sequencer_da_public_key: Vec<u8>,
+    /// The last processed l2 height in the processed sequencer commitments.
+    /// This will be 0 for pre fork 1 proofs
+    pub last_l2_height: u64,
+    /// Pre-proven commitments L2 ranges which also exist in the current L1 `da_data`.
+    pub preproven_commitments: Vec<usize>,
+}
