+++ crates/sovereign-sdk/rollup-interface/src/state_machine/zk/light_client_proof/mod.rs
@@ -1,0 +1,4 @@
+/// Light client proof input types and functions
+pub mod input;
+/// Light client proof output types and functions
+pub mod output;
